testing: avoid setting the source in the TestingSyncSource constructor
Calling createSource() should always be wrapped in CT_ macros. This
cannot be done when done as part of constructing a TestingSyncSource
on the stack, because then the constructed instance would be inside
the macro.

So let's remove the dead (and duplicated!) code. Revealed that the
code was still unintentionally used, fixed.