testing: avoid aborting client-test due to exception in destructor
The TestingSyncSource destructor ends up calling endSync() for
CalDAV/CardDAV, which must do some real work and may fail with an
exception. If the test has already failed and is getting left via
exception handling, then this extra work can be skipped and, more
importantly, no exceptions are allowed because they would terminate
client-test.

Happened in practice when Google becomes unresponsive with 401 errors,
because then the test fails and the reading of CTag in endSync() is
just going to encounter the same problem, thus raising a second
exception.