testing: added testLinkedItemsSubset for ActiveSync
This new test primarily covers the creation of artificial parent
events for detached recurrences in the activesyncd (BMC #22831). The
activesyncd must be run with EAS_DEBUG_DETACHED_RECURRENCES=1 set, so
that the artifical parent event is returned to the client. Otherwise
SyncEvolution cannot verify the correct creation of that event.

Putting this test into ClientTest.cpp instead of the ActiveSync
backend was done because these test cases may also serve as a useful
stress test for other backends. At the moment, they are only enabled
when CLIENT_TEST_SERVER is "exchange" or "apple". The artificial event
is only expected for "exchange".

A full test run with all imaginable combination of events would be
very slow. Therefore the tests choose roughly five starting points
scattered over the year 2012 and limit the number of events for each
test to five, except for the initial sequence with no events skipped:
that stores all events.

The the moment the test is known to fail with Exchange for those
weekly events which include the Sundays of the summer/winter time
change. The server uses incorrect UTC times in the UNTIL
clause. Perhaps the time zone definition that we send to it is not
quite correct?