utility: generic Init and InitState classes
The purpose of the older Bool class was to a) initialize the
underlying boolean when used as class member and b) to track whether
the value was ever set explicitly.

This patch adds the same functionality for all scalars (default
initialized with the help of Boost::value_init) and classes.

The user has to choose manually between InitState (for scalars) and
InitStateClass (for classes). The advantage of InitStateClass is that
it is derived from the base class, which allows easier access to the
base class methods: code which expected the base class still works.

For scalars it is sometimes necessary to cast into the base type or
use the get() method.

It would be nice if template specialization could be used to expose
just one InitState class. Boost::type_traits has a is_scalar member,
but all examples are about using that for specialization of functions
and it wasn't immediately obvious how that would extend to classes.
Perhaps sometime later inspiration will strike...