syncevo-http-server: support chained SSL certificates
So far, the file pointed to by --certificate-file had to
contain the server certificated (signed by a CA known to the client)
and (optionally) a client certificate.

With this patch, the file may also contain additional intermediate
certificates which will be sent to the client (chained certificates).

See http://twistedmatrix.com/pipermail/twisted-python/2010-July/022597.html