glib: avoid including glib/*.h headers directly
Recent glib deprecates the direct inclusion of some of its headers,
in favor of including glib.h. Doing that here whenever possible. gi18n.h
still needs to be included directly (otherwise it doesn't compile
on Debian Testing with glib 2.28).

Not sure about gio. I'm keeping the gio/ includes because they
are relevant on platforms where gio was not yet part of glib.