testing: run syncevo-dbus-server under valgrind (BMC #5609)
The test prefix needs to be passed through test-dbus.py to the
invocation of syncevo-dbus-server. Done with the new
TEST_DBUS_PREFIX env variable: set by runtests.py, used by test-dbus.py.

Also cleaned up the delay loop for starting syncevo-dbus-server:
the logic for reloading the log file was overly conservative. The file
only needs to be re-read when the file size has changed since the
last check (in contrast to the previous behavior or checkings always
if different from before the loop).