syncevo-dbus-server: fixed memory leaks in Bluetooth database manager
There was both a one-time memory leak (key file instance not freed)
and a per-method call memory leak (looking up values in the key file
returns allocated strings).

Fixed with GStringPtr utility class which takes over ownership of the
strings and by adding constructor+destructor to the owner of the key
file instance. Also mark the class as non-copyable, just in case.