D-Bus testing: more fixes for parsing of Python 2.7 output + HTML output
Some tests were incorrectly marked as "okay" because the there was
a line in the error message matching the "okay" regex. Now the pattern
also expects "okay" or "ok" at the end of the line.

The logic for detecting the start and end of an error report did not handle
multiple errors for a test. Now it does not overwrite an already opened
.log file anymore and keeps writing into it until some line really
indicates the beginning of another error.

As a new feature, the .html version of the .log file is written at the
same time as the .log file. It contains minor color coding. Its main
advantage is separation of D-Bus and stdout output and anchors with quick
links at the top of the page (can be used for skipping the D-Bus output).