D-Bus testing: further fixes for running under valgrind
The timeout of 5 seconds before hard killing the daemon was
too low when running it under valgrind. valgrind's leak checking
may take longer than that. Increased to one minute.

The timeout also sometimes was too low without valgrind. Increased
to 10 seconds.

Killing valgrindcheck.sh + syncevo-dbus-server is now more reliable
thanks to creating and killing a separate process group. Previously
a fatal SIGKILL was only delivered to valgrindcheck.sh and syncevo-dbus-server
kept running, breaking the following tests because the daemon
couldn't register itself with D-Bus.

Also ensure that the process group always gets a SIGKILL, even if the
main process we were waiting for has quit. Nightly testing sometimes
got stuck because a forked (?) process was still running. This might
point to a problem with process handling, but as that is going to change
anyway (fork+exec), don't bother now.