UUID: use rand_r()
Use rand() instead of the global rand(). This is motivated by two points: first,
rand() influences global state of the process, which is bad when done by
a library.

Second, valgrind (only once, though) reported:
==1994== Conditional jump or move depends on uninitialised value(s)
==1994==    at 0x8C3C7FC: srandom_r (random_r.c:198)
==1994==    by 0x8C3C490: srand (random.c:213)
==1994==    by 0x6037FE: SyncEvo::UUID::UUID() (util.cpp:322)

Note sure whether switching to rand_r() will change that, it still uses
the same way of obtaining the seed which was passed to srand().