testing: added testing of syncevo-http-server
The HTTP server and syncevo-dbus-server are started by two nested wrappercheck.sh
invocations. It would have been possible to use --start-syncevolution, but by
starting syncevo-dbus-server directly it is easier to check it with valgrind.

Both valgrindcheck.sh and wrappercheck.sh now support redirecting the output
of the command resp. daemon that they start, without having to also redirect
the output of the scripts themselves. This is used to redirect the output
of both syncevo-dbus-server and syncevo-http-server into syncevohttp.log,
from where client-test copies it into the *.server.log of the test.

wrappercheck.sh now also sends SIGTERM in addition to SIGINT.
Needed by Twisted on Debian Testing.