GLibSupport: added alloc function for g_spawn() string arrays
AllocStringArray() allocates at least the array of plain string
pointers and optionally also copies the strings. Arbitrary alloc/free
functions can be used. Default is to not copy strings and use new []
for the array. That matches a usage where the strings are owned by
some collection and the array is in a boost::scoped_array.