fork/exec: generic utility class for fork/exec + D-Bus
This depends on glib and D-Bus support. There is no global switch for
D-Bus on/off, only one for syncevo-dbus-server on/off, so that is what
is getting checked for now. The utility class is inside
libsyncevolution.la, so that library now also depends on
libgdbussyncevo.la, if enabled.

The actual implementation uses the GDBus client/server utility classes
abnd g_spawn(). A child watch is used to detect premature exits of the
child.

The idea was to also have a failure signal. Except for "child quit
without establishing connection" there is no error which currently has
to be reported, and that problem is already reported via the onQuit
signal. Therefore there is no onFailure signal yet.